Traverse BSP nodes from near to far
Traverse the BSP so that we first traverse nodes that are near to our
position and then traverse nodes that are far from it. The benefit for
us is reduced overdraw because Z-buffer is able to cull primitives that
are far away. It is also the BSP traversal order Doom uses.